Bug fixes; improvements to get<T>()

+++ CustomJSONData/CustomLevelInfo/CustomLevelInfoSaveData.cs
@@ -61,7 +61,7 @@ namespace CustomJSONData.CustomLevelInfo
                     return _customData;
                 }
             }
-            public DifficultyBeatmap(string difficultyName, int difficultyRank, string beatmapFilename, float noteJumpMovementSpeed, int noteJumpStartBeatOffset, dynamic customData) : base(difficultyName, difficultyRank, beatmapFilename, noteJumpMovementSpeed, noteJumpStartBeatOffset)
+            public DifficultyBeatmap(string difficultyName, int difficultyRank, string beatmapFilename, float noteJumpMovementSpeed, float noteJumpStartBeatOffset, dynamic customData) : base(difficultyName, difficultyRank, beatmapFilename, noteJumpMovementSpeed, noteJumpStartBeatOffset)
             {
                 _customData = customData;
             }
